% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-refcy3_8x60k_v1.R
\docType{data}
\name{refcy3_8x60k_v1}
\alias{refcy3_8x60k_v1}
\title{Cy3 Reference Intensities for Standard PBM 8x60k (v1) Design}
\description{
Reference intensities for computed for a set of 90 Cy3 scans
and was generated using \code{upbm::cy3GenerateRef}.
}
\seealso{
\code{\link[upbm]{cy3GenerateRef}}
}
\keyword{data}
